down read
big balance changes twice 22222

+++ Assets/Scripts/UnitStats.cs
@@ -6,11 +6,11 @@ using UnityEngine.UI;
 
 public class UnitStats : MonoBehaviour
 {
-    public float maxHP;
-    public float damage;
+    public int maxHP;
+    public int damage;
     public float attackSpeed; 
-    public float speed;
-    public float lifeTime;
+    public int speed;
+    public int lifeTime;
     public int cost;
     public int manacost;
     public int countOfUnits;
